requirements.txt
openai
python-dotenv
python-telegram-bot==12.7